pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "agentforge-backend"
version = "0.1.0"
description = "AgentForge Python/LangGraph backend for ClearForge AI"
requires-python = ">=3.11"

dependencies = [
    # Core agent framework
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.0",
    "langchain-anthropic>=0.1.0",

    # Memory
    "langmem>=0.0.15",

    # Tools / scraping
    "firecrawl-py>=0.0.16",
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.0",

    # CellCog SDK — minimum 1.12.0 required for latest feature support
    "cellcog>=1.12.0",

    # Telegram bot
    "python-telegram-bot>=21.0",

    # API server
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.29.0",

    # Supabase client
    "supabase>=2.4.0",

    # Utilities
    "python-dotenv>=1.0.0",
    "pydantic>=2.7.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.4.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true
